# backend_openapi_v3.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "NECO Photo Album API",
                 "description":  "FastAPI backend for NECO photo album generation",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/v1/uploads/dbf/state":  {
                                                    "post":  {
                                                                 "tags":  [
                                                                              "uploads"
                                                                          ],
                                                                 "summary":  "Upload State Dbf",
                                                                 "description":  "Upload state DBF file (step 1 of 3)",
                                                                 "operationId":  "upload_state_dbf_api_v1_uploads_dbf_state_post",
                                                                 "requestBody":  {
                                                                                     "content":  {
                                                                                                     "multipart/form-data":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/Body_upload_state_dbf_api_v1_uploads_dbf_state_post"
                                                                                                                                            }
                                                                                                                             }
                                                                                                 },
                                                                                     "required":  true
                                                                                 },
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "Successful Response",
                                                                                               "content":  {
                                                                                                               "application/json":  {
                                                                                                                                        "schema":  {

                                                                                                                                                   }
                                                                                                                                    }
                                                                                                           }
                                                                                           },
                                                                                   "422":  {
                                                                                               "description":  "Validation Error",
                                                                                               "content":  {
                                                                                                               "application/json":  {
                                                                                                                                        "schema":  {
                                                                                                                                                       "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                   }
                                                                                                                                    }
                                                                                                           }
                                                                                           }
                                                                               }
                                                             }
                                                },
                  "/api/v1/uploads/dbf/school":  {
                                                     "post":  {
                                                                  "tags":  [
                                                                               "uploads"
                                                                           ],
                                                                  "summary":  "Upload School Dbf",
                                                                  "description":  "Upload school DBF file (step 2 of 3)",
                                                                  "operationId":  "upload_school_dbf_api_v1_uploads_dbf_school_post",
                                                                  "requestBody":  {
                                                                                      "content":  {
                                                                                                      "multipart/form-data":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "$ref":  "#/components/schemas/Body_upload_school_dbf_api_v1_uploads_dbf_school_post"
                                                                                                                                             }
                                                                                                                              }
                                                                                                  },
                                                                                      "required":  true
                                                                                  },
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "Successful Response",
                                                                                                "content":  {
                                                                                                                "application/json":  {
                                                                                                                                         "schema":  {

                                                                                                                                                    }
                                                                                                                                     }
                                                                                                            }
                                                                                            },
                                                                                    "422":  {
                                                                                                "description":  "Validation Error",
                                                                                                "content":  {
                                                                                                                "application/json":  {
                                                                                                                                         "schema":  {
                                                                                                                                                        "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                    }
                                                                                                                                     }
                                                                                                            }
                                                                                            }
                                                                                }
                                                              }
                                                 },
                  "/api/v1/uploads/dbf/student":  {
                                                      "post":  {
                                                                   "tags":  [
                                                                                "uploads"
                                                                            ],
                                                                   "summary":  "Upload Student Dbf",
                                                                   "description":  "Upload student/master DBF file (step 3 of 3)",
                                                                   "operationId":  "upload_student_dbf_api_v1_uploads_dbf_student_post",
                                                                   "requestBody":  {
                                                                                       "content":  {
                                                                                                       "multipart/form-data":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "$ref":  "#/components/schemas/Body_upload_student_dbf_api_v1_uploads_dbf_student_post"
                                                                                                                                              }
                                                                                                                               }
                                                                                                   },
                                                                                       "required":  true
                                                                                   },
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Successful Response",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {

                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             },
                                                                                     "422":  {
                                                                                                 "description":  "Validation Error",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             }
                                                                                 }
                                                               }
                                                  },
                  "/api/v1/uploads/photos":  {
                                                 "post":  {
                                                              "tags":  [
                                                                           "uploads"
                                                                       ],
                                                              "summary":  "Upload Photos",
                                                              "operationId":  "upload_photos_api_v1_uploads_photos_post",
                                                              "requestBody":  {
                                                                                  "content":  {
                                                                                                  "multipart/form-data":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/Body_upload_photos_api_v1_uploads_photos_post"
                                                                                                                                         }
                                                                                                                          }
                                                                                              },
                                                                                  "required":  true
                                                                              },
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "Successful Response",
                                                                                            "content":  {
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {

                                                                                                                                                }
                                                                                                                                 }
                                                                                                        }
                                                                                        },
                                                                                "422":  {
                                                                                            "description":  "Validation Error",
                                                                                            "content":  {
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                }
                                                                                                                                 }
                                                                                                        }
                                                                                        }
                                                                            }
                                                          }
                                             },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "Body_upload_state_dbf_api_v1_uploads_dbf_state_post":  {
                                                                                                   "properties":  {
                                                                                                                      "state_dbf":  {
                                                                                                                                        "type":  "string",
                                                                                                                                        "format":  "binary",
                                                                                                                                        "title":  "State Dbf"
                                                                                                                                    }
                                                                                                                  },
                                                                                                   "type":  "object",
                                                                                                   "required":  [
                                                                                                                    "state_dbf"
                                                                                                                ],
                                                                                                   "title":  "Body_upload_state_dbf_api_v1_uploads_dbf_state_post"
                                                                                               },
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                   }
                   }
}